requirements.txt
python-telegram-bot>=20.0
matplotlib>=3.5.0
